chore: move preprocessing stuff to separate package
