+++ wonnx-preprocessing/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "wonnx-preprocessing"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+image = "0.23.14"
+log = "0.4.14"
+ndarray = "0.15.4"
+protobuf = { version = "2.25.2", features = ["with-bytes"] }
+thiserror = "1.0.30"
+tokenizers = "0.11.0"
+tract-onnx = { version = "0.15.8", optional = true }
+wgpu = "*"
+wonnx = { path = "../wonnx" }
