+++ wonnx-preprocessing/src/lib.rs
@@ -1,0 +1,10 @@
+use ndarray::ArrayBase;
+use wonnx::utils::Shape;
+
+pub mod image;
+pub mod text;
+
+pub struct Tensor {
+    pub data: ArrayBase<ndarray::OwnedRepr<f32>, ndarray::IxDyn>,
+    pub shape: Shape,
+}
