+++ wonnx/Cargo.toml
@@ -23,8 +23,6 @@ exclude = [
 [dependencies]
 wgpu = "0.12.0"
 bytemuck = "1.7.2"
-env_logger = "0.9.0"
-pollster = "0.2.4"
 log = "0.4.14" 
 protobuf = { version = "2", features = ["with-bytes"] }
 tera = "1.15.0"
@@ -38,6 +36,8 @@ num = "0.4.0"
 image = "0.23.14"
 ndarray = "0.15.4"
 approx = "0.5.0"
+pollster = "0.2.4"
+env_logger = "0.9.0"
 
 [target.'cfg(target_arch = "wasm32")'.dev-dependencies]
 console_error_panic_hook = "0.1.6"
